+++ test/FeesHandler.t.sol
@@ -470,7 +470,7 @@ contract TestableFeesHandler is Depositor, FeesHandler {
 
   // -------- Overridden abstract function placeholders --------
 
-  function convertToReceiptTokenAmount(uint256, /* reservePoolId_ */ uint256 /*reserveAssetAmount_ */ )
+  function convertToReceiptTokenAmount(uint8, /* reservePoolId_ */ uint256 /*reserveAssetAmount_ */ )
     public
     view
     override
@@ -479,7 +479,7 @@ contract TestableFeesHandler is Depositor, FeesHandler {
     __readStub__();
   }
 
-  function convertToReserveAssetAmount(uint256, /* depositReceiptTokenAmount_ */ uint256 /* reservePoolId_ */ )
+  function convertToReserveAssetAmount(uint8, /* reservePoolId_ */ uint256 /* depositReceiptTokenAmount_ */ )
     public
     view
     override
